moved pentagon to new module
