+++ drawing/src/canvas2d/mod.rs
@@ -172,189 +172,7 @@ pub struct Canvas2D {
     camera: Box<dyn Camera>,
 }
 
-#[repr(C)]
-#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
-struct Vertex {
-    position: [f32; 2],
-    color: [f32; 3],
-}
-
-impl Vertex {
-    const ATTRIBS: [wgpu::VertexAttribute; 2] =
-        wgpu::vertex_attr_array![0 => Float32x2, 1 => Float32x3];
-
-    fn desc() -> wgpu::VertexBufferLayout<'static> {
-        wgpu::VertexBufferLayout {
-            array_stride: std::mem::size_of::<Vertex>() as wgpu::BufferAddress,
-            step_mode: wgpu::VertexStepMode::Vertex,
-            attributes: &Self::ATTRIBS,
-        }
-    }
-}
-
-const VERTICES: &[Vertex] = &[
-    Vertex {
-        position: [-8.68241, 49.240386],
-        color: [0.5, 0.0, 0.5],
-    }, // A
-    Vertex {
-        position: [-49.513406, 6.958647],
-        color: [0.5, 0.5, 0.5],
-    }, // B
-    Vertex {
-        position: [-21.918549, -44.939706],
-        color: [0.5, 0.0, 0.5],
-    }, // C
-    Vertex {
-        position: [35.966998, -34.73291],
-        color: [0.5, 0.0, 0.5],
-    }, // D
-    Vertex {
-        position: [44.147372, 23.47359],
-        color: [0.5, 0.0, 0.5],
-    }, // E
-];
-
-const INDICES: &[u16] = &[0, 1, 4, 1, 2, 4, 2, 3, 4];
-
-pub struct PentagonWgpu {
-    vertex_buffer: wgpu::Buffer,
-    num_vertices: u32,
-    index_buffer: wgpu::Buffer,
-    num_indices: u32,
-    render_pipeline: wgpu::RenderPipeline,
-}
-
-impl PentagonWgpu {
-    fn new(wgpu_state: &WgpuState, camera_bind_group_layout: &BindGroupLayout) -> Self {
-        let vertex_buffer =
-            wgpu_state
-                .device
-                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                    label: Some("Vertex Buffer"),
-                    contents: bytemuck::cast_slice(VERTICES),
-                    usage: wgpu::BufferUsages::VERTEX,
-                });
-
-        let num_vertices = VERTICES.len() as u32;
-
-        let index_buffer =
-            wgpu_state
-                .device
-                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                    label: Some("Index Buffer"),
-                    contents: bytemuck::cast_slice(INDICES),
-                    usage: wgpu::BufferUsages::INDEX,
-                });
-
-        let num_indices = INDICES.len() as u32;
-
-        let shader = wgpu_state
-            .device
-            .create_shader_module(wgpu::ShaderModuleDescriptor {
-                label: Some("Shader"),
-                source: wgpu::ShaderSource::Wgsl(include_str!("shader.wgsl").into()),
-            });
-
-        let render_pipeline_layout =
-            wgpu_state
-                .device
-                .create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-                    label: Some("Render Pipeline Layout"),
-                    bind_group_layouts: &[camera_bind_group_layout],
-                    push_constant_ranges: &[],
-                });
-
-        let render_pipeline =
-            wgpu_state
-                .device
-                .create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-                    label: Some("Render Pipeline"),
-                    layout: Some(&render_pipeline_layout),
-                    vertex: wgpu::VertexState {
-                        module: &shader,
-                        entry_point: Some("vs_main"),
-                        buffers: &[Vertex::desc()],
-                        compilation_options: wgpu::PipelineCompilationOptions::default(),
-                    },
-                    fragment: Some(wgpu::FragmentState {
-                        // 3.
-                        module: &shader,
-                        entry_point: Some("fs_main"),
-                        targets: &[Some(wgpu::ColorTargetState {
-                            // 4.
-                            format: wgpu_state.config.format,
-                            blend: Some(wgpu::BlendState::REPLACE),
-                            write_mask: wgpu::ColorWrites::ALL,
-                        })],
-                        compilation_options: wgpu::PipelineCompilationOptions::default(),
-                    }),
-                    primitive: wgpu::PrimitiveState {
-                        topology: wgpu::PrimitiveTopology::TriangleList,
-                        strip_index_format: None,
-                        front_face: wgpu::FrontFace::Ccw,
-                        cull_mode: Some(wgpu::Face::Back),
-                        // Setting this to anything other than Fill requires Features::NON_FILL_POLYGON_MODE
-                        polygon_mode: wgpu::PolygonMode::Fill,
-                        // Requires Features::DEPTH_CLIP_CONTROL
-                        unclipped_depth: false,
-                        // Requires Features::CONSERVATIVE_RASTERIZATION
-                        conservative: false,
-                    },
-                    depth_stencil: None,
-                    multisample: wgpu::MultisampleState {
-                        count: 1,
-                        mask: !0,
-                        alpha_to_coverage_enabled: false,
-                    },
-                    multiview: None,
-                    cache: None,
-                });
-
-        Self {
-            vertex_buffer,
-            num_vertices,
-            index_buffer,
-            num_indices,
-            render_pipeline,
-        }
-    }
-
-    fn render(
-        &mut self,
-        encoder: &mut CommandEncoder,
-        view: &TextureView,
-        camera_bind_group: &BindGroup,
-    ) -> Result<(), wgpu::SurfaceError> {
-        let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-            label: Some("Render Pass"),
-            color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                view: view,
-                resolve_target: None,
-                ops: wgpu::Operations {
-                    load: wgpu::LoadOp::Clear(wgpu::Color {
-                        r: 0.1,
-                        g: 0.2,
-                        b: 0.3,
-                        a: 1.0,
-                    }),
-                    store: wgpu::StoreOp::Store,
-                },
-            })],
-            depth_stencil_attachment: None,
-            occlusion_query_set: None,
-            timestamp_writes: None,
-        });
-
-        render_pass.set_pipeline(&self.render_pipeline);
-        render_pass.set_bind_group(0, camera_bind_group, &[]);
-        render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
-        render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
-
-        render_pass.draw_indexed(0..self.num_indices, 0, 0..1);
-        Ok(())
-    }
-}
+mod pentagon;
 
 pub struct Canvas2DWindowState {
     wgpu_state: WgpuState,
@@ -363,7 +181,7 @@ pub struct Canvas2DWindowState {
     camera_buffer: wgpu::Buffer,
     camera_bind_group: wgpu::BindGroup,
 
-    pentagon: PentagonWgpu,
+    pentagon: pentagon::PentagonWgpu,
 }
 
 impl Canvas2DWindowState {
@@ -413,7 +231,7 @@ impl Canvas2DWindowState {
             });
 
         Self {
-            pentagon: PentagonWgpu::new(&wgpu_state, &camera_bind_group_layout),
+            pentagon: pentagon::PentagonWgpu::new(&wgpu_state, &camera_bind_group_layout),
             wgpu_state,
             camera_uniform,
             camera_bind_group,
