+++ drawing/src/canvas2d/pentagon/shader.wgsl
